Added player distance and limits to sliding and jumping
The player can no longer slide while on an object (will need to modify sliding bounding box to allow for you to slide while on an object). Also, you can no longer get out of a slide via jumping. This is to prevent a player from jumping into an obstacle while under it. Instead, the player is forced to remain in the slide until the obstacle is no longer present.