Screenshots for data recording
When screenshots are enabled (toggle in constants.py) a screenshot is taken either every SCREENSHOT_FREQUENCY (change in constants.py) or when an action is taken (S is pressed or W is keyed down). Also, time elapsed is now shown on the hud and the player character automatically moves forward.